Remove special knowledge of ElevenLabs from ServiceMixin
